@@ -6,11 +6,12 @@ homepage = "https://github.com/sharkdp/bat"
 license = "MIT OR Apache-2.0"
 name = "bat"
 repository = "https://github.com/sharkdp/bat"
-version = "0.24.0"
+version = "0.25.0"
 exclude = ["assets/syntaxes/*", "assets/themes/*"]
 build = "build/main.rs"
 edition = '2021'
-rust-version = "1.70"
+# You are free to bump MSRV as soon as a reason for bumping emerges.
+rust-version = "1.74"
 
 [features]
 default = ["application"]
@@ -33,7 +34,7 @@ minimal-application = [
 ]
 git = ["git2"] # Support indicating git modifications
 paging = ["shell-words", "grep-cli"] # Support applying a pager on the output
-lessopen = ["run_script", "os_str_bytes"] # Support $LESSOPEN preprocessor
+lessopen = ["execute"] # Support $LESSOPEN preprocessor
 build-assets = ["syntect/yaml-load", "syntect/plist-load", "regex", "walkdir"]
 
 # You need to use one of these if you depend on bat as a library:
@@ -44,33 +45,35 @@ regex-fancy = ["syntect/regex-fancy"] # Use the rust-only "fancy-regex" engine
 nu-ansi-term = "0.50.0"
 ansi_colours = "^1.2"
 bincode = "1.0"
-console = "0.15.8"
-flate2 = "1.0"
-once_cell = "1.19"
-thiserror = "1.0"
+console = "0.16.0"
+flate2 = "1.1"
+once_cell = "1.20"
+thiserror = "2.0"
 wild = { version = "2.2", optional = true }
 content_inspector = "0.2.4"
 shell-words = { version = "1.1.0", optional = true }
-unicode-width = "0.1.11"
+unicode-width = "0.2.0"
 globset = "0.4"
 serde = "1.0"
 serde_derive = "1.0"
 serde_yaml = "0.9.28"
 semver = "1.0"
 path_abs = { version = "0.5", default-features = false }
-clircle = "0.5"
+clircle = { version = "0.6.1", default-features = false }
 bugreport = { version = "0.5.0", optional = true }
-etcetera = { version = "0.8.0", optional = true }
-grep-cli = { version = "0.1.10", optional = true }
-regex = { version = "1.10.2", optional = true }
-walkdir = { version = "2.4", optional = true }
+etcetera = { version = "0.10.0", optional = true }
+grep-cli = { version = "0.1.11", optional = true }
+regex = { version = "1.10.6", optional = true }
+walkdir = { version = "2.5", optional = true }
 bytesize = { version = "1.3.0" }
-encoding_rs = "0.8.33"
-os_str_bytes = { version = "~7.0", optional = true }
-run_script = { version = "^0.10.1", optional = true}
+encoding_rs = "0.8.35"
+execute = { version = "0.2.13", optional = true }
+terminal-colorsaurus = "1.0"
+unicode-segmentation = "1.12.0"
+itertools = "0.13.0"
 
 [dependencies.git2]
-version = "0.18"
+version = "0.20"
 optional = true
 default-features = false
 
@@ -86,31 +89,35 @@ features = ["wrap_help", "cargo"]
 
 [target.'cfg(target_os = "macos")'.dependencies]
 home = "0.5.9"
-plist = "1.6.0"
+plist = "1.7.0"
 
 [dev-dependencies]
 assert_cmd = "2.0.12"
-expect-test = "1.4.1"
+expect-test = "1.5.0"
 serial_test = { version = "2.0.0", default-features = false }
-predicates = "3.1.0"
+predicates = "3.1.3"
 wait-timeout = "0.2.0"
-tempfile = "3.8.1"
+tempfile = "3.16.0"
 serde = { version = "1.0", features = ["derive"] }
 
 [target.'cfg(unix)'.dev-dependencies]
-nix = { version = "0.26.4", default-features = false, features = ["term"] }
+nix = { version = "0.30", default-features = false, features = ["term"] }
 
 [build-dependencies]
-anyhow = "1.0.78"
-indexmap = { version = "2.2.6", features = ["serde"] }
-itertools = "0.13.0"
-once_cell = "1.18"
-regex = "1.10.2"
+anyhow = "1.0.97"
+indexmap = { version = "2.8.0", features = ["serde"] }
+itertools = "0.14.0"
+once_cell = "1.20"
+prettyplease = "0.2.35"
+proc-macro2 = "1.0.95"
+quote = "1.0.40"
+regex = "1.10.6"
 serde = "1.0"
 serde_derive = "1.0"
-serde_with = { version = "3.8.1", default-features = false, features = ["macros"] }
-toml = { version = "0.8.9", features = ["preserve_order"] }
-walkdir = "2.4"
+serde_with = { version = "3.12.0", default-features = false, features = ["macros"] }
+syn = { version = "2.0.104", features = ["full"] }
+toml = { version = "0.9.1", features = ["preserve_order"] }
+walkdir = "2.5"
 
 [build-dependencies.clap]
 version = "4.4.12"
