@@ -75,7 +75,7 @@ pub(crate) enum InputKind<'a> {
     CustomReader(Box<dyn Read + 'a>),
 }
 
-impl<'a> InputKind<'a> {
+impl InputKind<'_> {
     pub fn description(&self) -> InputDescription {
         match self {
             InputKind::OrdinaryFile(ref path) => InputDescription::new(path.to_string_lossy()),
@@ -217,14 +217,14 @@ impl<'a> Input<'a> {
                 metadata: self.metadata,
                 reader: {
                     let mut file = File::open(&path)
-                        .map_err(|e| format!("'{}': {}", path.to_string_lossy(), e))?;
+                        .map_err(|e| format!("'{}': {e}", path.to_string_lossy()))?;
                     if file.metadata()?.is_dir() {
                         return Err(format!("'{}' is a directory.", path.to_string_lossy()).into());
                     }
 
                     if let Some(stdout) = stdout_identifier {
                         let input_identifier = Identifier::try_from(file).map_err(|e| {
-                            format!("{}: Error identifying file: {}", path.to_string_lossy(), e)
+                            format!("{}: Error identifying file: {e}", path.to_string_lossy())
                         })?;
                         if stdout.surely_conflicts_with(&input_identifier) {
                             return Err(format!(
@@ -267,7 +267,9 @@ impl<'a> InputReader<'a> {
         };
 
         if content_type == Some(ContentType::UTF_16LE) {
-            reader.read_until(0x00, &mut first_line).ok();
+            read_utf16_line(&mut reader, &mut first_line, 0x00, 0x0A).ok();
+        } else if content_type == Some(ContentType::UTF_16BE) {
+            read_utf16_line(&mut reader, &mut first_line, 0x0A, 0x00).ok();
         }
 
         InputReader {
@@ -283,14 +285,42 @@ impl<'a> InputReader<'a> {
             return Ok(true);
         }
 
-        let res = self.inner.read_until(b'\n', buf).map(|size| size > 0)?;
-
         if self.content_type == Some(ContentType::UTF_16LE) {
-            let _ = self.inner.read_until(0x00, buf);
+            return read_utf16_line(&mut self.inner, buf, 0x00, 0x0A);
+        }
+        if self.content_type == Some(ContentType::UTF_16BE) {
+            return read_utf16_line(&mut self.inner, buf, 0x0A, 0x00);
         }
 
+        let res = self.inner.read_until(b'\n', buf).map(|size| size > 0)?;
         Ok(res)
     }
+}
+
+fn read_utf16_line<R: BufRead>(
+    reader: &mut R,
+    buf: &mut Vec<u8>,
+    read_until_char: u8,
+    preceded_by_char: u8,
+) -> io::Result<bool> {
+    loop {
+        let mut temp = Vec::new();
+        let n = reader.read_until(read_until_char, &mut temp)?;
+        if n == 0 {
+            // EOF reached
+            break;
+        }
+        buf.extend_from_slice(&temp);
+        if buf.len() >= 2
+            && buf[buf.len() - 2] == preceded_by_char
+            && buf[buf.len() - 1] == read_until_char
+        {
+            // end of line found
+            break;
+        }
+        // end of line not found, keep going
+    }
+    Ok(!buf.is_empty())
 }
 
 #[test]
@@ -342,6 +372,56 @@ fn utf16le() {
     assert!(res.is_ok());
     assert!(res.unwrap());
     assert_eq!(b"\x64\x00", &buffer[..]);
+
+    buffer.clear();
+
+    let res = reader.read_line(&mut buffer);
+    assert!(res.is_ok());
+    assert!(!res.unwrap());
+    assert!(buffer.is_empty());
+}
+
+#[test]
+fn utf16le_issue3367() {
+    let content = b"\xFF\xFE\x0A\x4E\x00\x4E\x0A\x4F\x00\x52\x0A\x00\
+        \x6F\x00\x20\x00\x62\x00\x61\x00\x72\x00\x0A\x00\
+        \x68\x00\x65\x00\x6C\x00\x6C\x00\x6F\x00\x20\x00\x77\x00\x6F\x00\x72\x00\x6C\x00\x64\x00";
+    let mut reader = InputReader::new(&content[..]);
+
+    assert_eq!(
+        b"\xFF\xFE\x0A\x4E\x00\x4E\x0A\x4F\x00\x52\x0A\x00",
+        &reader.first_line[..]
+    );
+
+    let mut buffer = vec![];
+
+    let res = reader.read_line(&mut buffer);
+    assert!(res.is_ok());
+    assert!(res.unwrap());
+    assert_eq!(
+        b"\xFF\xFE\x0A\x4E\x00\x4E\x0A\x4F\x00\x52\x0A\x00",
+        &buffer[..]
+    );
+
+    buffer.clear();
+
+    let res = reader.read_line(&mut buffer);
+    assert!(res.is_ok());
+    assert!(res.unwrap());
+    assert_eq!(
+        b"\x6F\x00\x20\x00\x62\x00\x61\x00\x72\x00\x0A\x00",
+        &buffer[..]
+    );
+
+    buffer.clear();
+
+    let res = reader.read_line(&mut buffer);
+    assert!(res.is_ok());
+    assert!(res.unwrap());
+    assert_eq!(
+        b"\x68\x00\x65\x00\x6C\x00\x6C\x00\x6F\x00\x20\x00\x77\x00\x6F\x00\x72\x00\x6C\x00\x64\x00",
+        &buffer[..]
+    );
 
     buffer.clear();
 
