@@ -1,5 +1,3 @@
-use std::fmt;
-use std::io;
 use std::vec::Vec;
 
 use nu_ansi_term::Color::{Fixed, Green, Red, Yellow};
@@ -17,6 +15,7 @@ use content_inspector::ContentType;
 
 use encoding_rs::{UTF_16BE, UTF_16LE};
 
+use unicode_segmentation::UnicodeSegmentation;
 use unicode_width::UnicodeWidthChar;
 
 use crate::assets::{HighlightingAssets, SyntaxReferenceInSet};
@@ -28,13 +27,15 @@ use crate::decorations::{Decoration, GridBorderDecoration, LineNumberDecoration}
 use crate::diff::LineChanges;
 use crate::error::*;
 use crate::input::OpenedInput;
-use crate::line_range::RangeCheckResult;
+use crate::line_range::{MaxBufferedLineNumber, RangeCheckResult};
+use crate::output::OutputHandle;
 use crate::preprocessor::strip_ansi;
 use crate::preprocessor::{expand_tabs, replace_nonprintable};
 use crate::style::StyleComponent;
 use crate::terminal::{as_terminal_escaped, to_ansi_color};
 use crate::vscreen::{AnsiStyle, EscapeSequence, EscapeSequenceIterator};
 use crate::wrapping::WrappingMode;
+use crate::BinaryBehavior;
 use crate::StripAnsiMode;
 
 const ANSI_UNDERLINE_ENABLE: EscapeSequence = EscapeSequence::CSI {
@@ -67,20 +68,6 @@ const EMPTY_SYNTECT_STYLE: syntect::highlighting::Style = syntect::highlighting:
     font_style: FontStyle::empty(),
 };
 
-pub enum OutputHandle<'a> {
-    IoWrite(&'a mut dyn io::Write),
-    FmtWrite(&'a mut dyn fmt::Write),
-}
-
-impl<'a> OutputHandle<'a> {
-    fn write_fmt(&mut self, args: fmt::Arguments<'_>) -> Result<()> {
-        match self {
-            Self::IoWrite(handle) => handle.write_fmt(args).map_err(Into::into),
-            Self::FmtWrite(handle) => handle.write_fmt(args).map_err(Into::into),
-        }
-    }
-}
-
 pub(crate) trait Printer {
     fn print_header(
         &mut self,
@@ -98,6 +85,7 @@ pub(crate) trait Printer {
         handle: &mut OutputHandle,
         line_number: usize,
         line_buffer: &[u8],
+        max_buffered_line_number: MaxBufferedLineNumber,
     ) -> Result<()>;
 }
 
@@ -115,7 +103,7 @@ impl<'a> SimplePrinter<'a> {
     }
 }
 
-impl<'a> Printer for SimplePrinter<'a> {
+impl Printer for SimplePrinter<'_> {
     fn print_header(
         &mut self,
         _handle: &mut OutputHandle,
@@ -139,11 +127,12 @@ impl<'a> Printer for SimplePrinter<'a> {
         handle: &mut OutputHandle,
         _line_number: usize,
         line_buffer: &[u8],
+        _max_buffered_line_number: MaxBufferedLineNumber,
     ) -> Result<()> {
         // Skip squeezed lines.
         if let Some(squeeze_limit) = self.config.squeeze_lines {
             if String::from_utf8_lossy(line_buffer)
-                .trim_end_matches(|c| c == '\r' || c == '\n')
+                .trim_end_matches(['\r', '\n'])
                 .is_empty()
             {
                 self.consecutive_empty_lines += 1;
@@ -266,9 +255,10 @@ impl<'a> InteractivePrinter<'a> {
         let is_printing_binary = input
             .reader
             .content_type
-            .map_or(false, |c| c.is_binary() && !config.show_nonprintable);
+            .is_some_and(|c| c.is_binary() && !config.show_nonprintable);
 
-        let needs_to_match_syntax = !is_printing_binary
+        let needs_to_match_syntax = (!is_printing_binary
+            || matches!(config.binary, BinaryBehavior::AsText))
             && (config.colored_output || config.strip_ansi == StripAnsiMode::Auto);
 
         let (is_plain_text, highlighter_from_set) = if needs_to_match_syntax {
@@ -339,7 +329,7 @@ impl<'a> InteractivePrinter<'a> {
             self.print_horizontal_line_term(handle, self.colors.grid)?;
         } else {
             let hline = "─".repeat(self.config.term_width - (self.panel_width + 1));
-            let hline = format!("{}{}{}", "─".repeat(self.panel_width), grid_char, hline);
+            let hline = format!("{}{grid_char}{hline}", "─".repeat(self.panel_width));
             writeln!(handle, "{}", self.colors.grid.paint(hline))?;
         }
 
@@ -353,8 +343,7 @@ impl<'a> InteractivePrinter<'a> {
 
         let text_truncated: String = text.chars().take(self.panel_width - 1).collect();
         let text_filled: String = format!(
-            "{}{}",
-            text_truncated,
+            "{text_truncated}{}",
             " ".repeat(self.panel_width - 1 - text_truncated.len())
         );
         if self.config.style_components.grid() {
@@ -401,14 +390,18 @@ impl<'a> InteractivePrinter<'a> {
         handle: &mut OutputHandle,
         content: &str,
     ) -> Result<()> {
-        let mut content = content;
         let content_width = self.config.term_width - self.get_header_component_indent_length();
-        while content.len() > content_width {
-            let (content_line, remaining) = content.split_at(content_width);
-            self.print_header_component_with_indent(handle, content_line)?;
-            content = remaining;
+        if content.chars().count() <= content_width {
+            return self.print_header_component_with_indent(handle, content);
         }
-        self.print_header_component_with_indent(handle, content)
+
+        let mut content_graphemes: Vec<&str> = content.graphemes(true).collect();
+        while content_graphemes.len() > content_width {
+            let (content_line, remaining) = content_graphemes.split_at(content_width);
+            self.print_header_component_with_indent(handle, content_line.join("").as_str())?;
+            content_graphemes = remaining.to_vec();
+        }
+        self.print_header_component_with_indent(handle, content_graphemes.join("").as_str())
     }
 
     fn highlight_regions_for_line<'b>(
@@ -430,7 +423,7 @@ impl<'a> InteractivePrinter<'a> {
             .highlight_line(for_highlighting, highlighter_from_set.syntax_set)?;
 
         if too_long {
-            highlighted_line[0].1 = &line;
+            highlighted_line[0].1 = line;
         }
 
         Ok(highlighted_line)
@@ -446,7 +439,7 @@ impl<'a> InteractivePrinter<'a> {
     }
 }
 
-impl<'a> Printer for InteractivePrinter<'a> {
+impl Printer for InteractivePrinter<'_> {
     fn print_header(
         &mut self,
         handle: &mut OutputHandle,
@@ -458,7 +451,10 @@ impl<'a> Printer for InteractivePrinter<'a> {
         }
 
         if !self.config.style_components.header() {
-            if Some(ContentType::BINARY) == self.content_type && !self.config.show_nonprintable {
+            if Some(ContentType::BINARY) == self.content_type
+                && !self.config.show_nonprintable
+                && !matches!(self.config.binary, BinaryBehavior::AsText)
+            {
                 writeln!(
                     handle,
                     "{}: Binary content from {} will not be printed to the terminal \
@@ -516,13 +512,12 @@ impl<'a> Printer for InteractivePrinter<'a> {
             .try_for_each(|component| match component {
                 StyleComponent::HeaderFilename => {
                     let header_filename = format!(
-                        "{}{}{}",
+                        "{}{}{mode}",
                         description
                             .kind()
                             .map(|kind| format!("{kind}: "))
                             .unwrap_or_else(|| "".into()),
                         self.colors.header_value.paint(description.title()),
-                        mode
                     );
                     self.print_header_multiline_component(handle, &header_filename)
                 }
@@ -539,7 +534,10 @@ impl<'a> Printer for InteractivePrinter<'a> {
             })?;
 
         if self.config.style_components.grid() {
-            if self.content_type.map_or(false, |c| c.is_text()) || self.config.show_nonprintable {
+            if self.content_type.is_some_and(|c| c.is_text())
+                || self.config.show_nonprintable
+                || matches!(self.config.binary, BinaryBehavior::AsText)
+            {
                 self.print_horizontal_line(handle, '┼')?;
             } else {
                 self.print_horizontal_line(handle, '┴')?;
@@ -551,7 +549,9 @@ impl<'a> Printer for InteractivePrinter<'a> {
 
     fn print_footer(&mut self, handle: &mut OutputHandle, _input: &OpenedInput) -> Result<()> {
         if self.config.style_components.grid()
-            && (self.content_type.map_or(false, |c| c.is_text()) || self.config.show_nonprintable)
+            && (self.content_type.is_some_and(|c| c.is_text())
+                || self.config.show_nonprintable
+                || matches!(self.config.binary, BinaryBehavior::AsText))
         {
             self.print_horizontal_line(handle, '┴')
         } else {
@@ -589,6 +589,7 @@ impl<'a> Printer for InteractivePrinter<'a> {
         handle: &mut OutputHandle,
         line_number: usize,
         line_buffer: &[u8],
+        max_buffered_line_number: MaxBufferedLineNumber,
     ) -> Result<()> {
         let line = if self.config.show_nonprintable {
             replace_nonprintable(
@@ -599,7 +600,9 @@ impl<'a> Printer for InteractivePrinter<'a> {
             .into()
         } else {
             let mut line = match self.content_type {
-                Some(ContentType::BINARY) | None => {
+                Some(ContentType::BINARY) | None
+                    if !matches!(self.config.binary, BinaryBehavior::AsText) =>
+                {
                     return Ok(());
                 }
                 Some(ContentType::UTF_16LE) => UTF_16LE.decode_with_bom_removal(line_buffer).0,
@@ -632,7 +635,7 @@ impl<'a> Printer for InteractivePrinter<'a> {
 
         // Skip squeezed lines.
         if let Some(squeeze_limit) = self.config.squeeze_lines {
-            if line.trim_end_matches(|c| c == '\r' || c == '\n').is_empty() {
+            if line.trim_end_matches(['\r', '\n']).is_empty() {
                 self.consecutive_empty_lines += 1;
                 if self.consecutive_empty_lines > squeeze_limit {
                     return Ok(());
@@ -648,8 +651,12 @@ impl<'a> Printer for InteractivePrinter<'a> {
         let mut panel_wrap: Option<String> = None;
 
         // Line highlighting
-        let highlight_this_line =
-            self.config.highlighted_lines.0.check(line_number) == RangeCheckResult::InRange;
+        let highlight_this_line = self
+            .config
+            .highlighted_lines
+            .0
+            .check(line_number, max_buffered_line_number)
+            == RangeCheckResult::InRange;
 
         if highlight_this_line && self.config.theme == "ansi" {
             self.ansi_style.update(ANSI_UNDERLINE_ENABLE);
@@ -685,14 +692,14 @@ impl<'a> Printer for InteractivePrinter<'a> {
                         // Regular text.
                         EscapeSequence::Text(text) => {
                             let text = self.preprocess(text, &mut cursor_total);
-                            let text_trimmed = text.trim_end_matches(|c| c == '\r' || c == '\n');
+                            let text_trimmed = text.trim_end_matches(['\r', '\n']);
 
                             write!(
                                 handle,
                                 "{}{}",
                                 as_terminal_escaped(
                                     style,
-                                    &format!("{}{}", self.ansi_style, text_trimmed),
+                                    &format!("{}{text_trimmed}", self.ansi_style),
                                     true_color,
                                     colored_output,
                                     italics,
@@ -739,10 +746,8 @@ impl<'a> Printer for InteractivePrinter<'a> {
                     match chunk {
                         // Regular text.
                         EscapeSequence::Text(text) => {
-                            let text = self.preprocess(
-                                text.trim_end_matches(|c| c == '\r' || c == '\n'),
-                                &mut cursor_total,
-                            );
+                            let text = self
+                                .preprocess(text.trim_end_matches(['\r', '\n']), &mut cursor_total);
 
                             let mut max_width = cursor_max - cursor;
 
@@ -784,7 +789,7 @@ impl<'a> Printer for InteractivePrinter<'a> {
                                         "{}{}\n{}",
                                         as_terminal_escaped(
                                             style,
-                                            &format!("{}{}", self.ansi_style, line_buf),
+                                            &format!("{}{line_buf}", self.ansi_style),
                                             self.config.true_color,
                                             self.config.colored_output,
                                             self.config.use_italic_text,
@@ -811,7 +816,7 @@ impl<'a> Printer for InteractivePrinter<'a> {
                                 "{}",
                                 as_terminal_escaped(
                                     style,
-                                    &format!("{}{}", self.ansi_style, line_buf),
+                                    &format!("{}{line_buf}", self.ansi_style),
                                     self.config.true_color,
                                     self.config.colored_output,
                                     self.config.use_italic_text,
