@@ -1,4 +1,6 @@
-use bat::{assets::HighlightingAssets, config::Config, controller::Controller, Input};
+use bat::{
+    assets::HighlightingAssets, config::Config, controller::Controller, output::OutputHandle, Input,
+};
 
 fn main() {
     let mut buffer = String::new();
@@ -10,7 +12,10 @@ fn main() {
     let controller = Controller::new(&config, &assets);
     let input = Input::from_file(file!());
     controller
-        .run(vec![input.into()], Some(&mut buffer))
+        .run(
+            vec![input.into()],
+            Some(OutputHandle::FmtWrite(&mut buffer)),
+        )
         .unwrap();
 
     println!("{buffer}");
