@@ -13,6 +13,7 @@ use crate::error::*;
 use crate::input::{InputReader, OpenedInput};
 use crate::syntax_mapping::ignored_suffixes::IgnoredSuffixes;
 use crate::syntax_mapping::MappingTarget;
+use crate::theme::{default_theme, ColorScheme};
 use crate::{bat_warning, SyntaxMapping};
 
 use lazy_theme_set::LazyThemeSet;
@@ -67,57 +68,6 @@ impl HighlightingAssets {
             theme_set,
             fallback_theme: None,
         }
-    }
-
-    /// The default theme.
-    ///
-    /// ### Windows and Linux
-    ///
-    /// Windows and most Linux distributions has a dark terminal theme by
-    /// default. On these platforms, this function always returns a theme that
-    /// looks good on a dark background.
-    ///
-    /// ### macOS
-    ///
-    /// On macOS the default terminal background is light, but it is common that
-    /// Dark Mode is active, which makes the terminal background dark. On this
-    /// platform, the default theme depends on
-    /// ```bash
-    /// defaults read -globalDomain AppleInterfaceStyle
-    /// ```
-    /// To avoid the overhead of the check on macOS, simply specify a theme
-    /// explicitly via `--theme`, `BAT_THEME`, or `~/.config/bat`.
-    ///
-    /// See <https://github.com/sharkdp/bat/issues/1746> and
-    /// <https://github.com/sharkdp/bat/issues/1928> for more context.
-    pub fn default_theme() -> &'static str {
-        #[cfg(not(target_os = "macos"))]
-        {
-            Self::default_dark_theme()
-        }
-        #[cfg(target_os = "macos")]
-        {
-            if macos_dark_mode_active() {
-                Self::default_dark_theme()
-            } else {
-                Self::default_light_theme()
-            }
-        }
-    }
-
-    /**
-     * The default theme that looks good on a dark background.
-     */
-    fn default_dark_theme() -> &'static str {
-        "Monokai Extended"
-    }
-
-    /**
-     * The default theme that looks good on a light background.
-     */
-    #[cfg(target_os = "macos")]
-    fn default_light_theme() -> &'static str {
-        "Monokai Extended Light"
     }
 
     pub fn from_cache(cache_path: &Path) -> Result<Self> {
@@ -202,7 +152,7 @@ impl HighlightingAssets {
         &self,
         path: impl AsRef<Path>,
         mapping: &SyntaxMapping,
-    ) -> Result<SyntaxReferenceInSet> {
+    ) -> Result<SyntaxReferenceInSet<'_>> {
         let path = path.as_ref();
 
         let syntax_match = mapping.get_syntax_for(path);
@@ -213,7 +163,7 @@ impl HighlightingAssets {
 
         if let Some(MappingTarget::MapTo(syntax_name)) = syntax_match {
             return self
-                .find_syntax_by_name(syntax_name)?
+                .find_syntax_by_token(syntax_name)?
                 .ok_or_else(|| Error::UnknownSyntax(syntax_name.to_owned()));
         }
 
@@ -241,14 +191,17 @@ impl HighlightingAssets {
             Some(theme) => theme,
             None => {
                 if theme == "ansi-light" || theme == "ansi-dark" {
-                    bat_warning!("Theme '{}' is deprecated, using 'ansi' instead.", theme);
+                    bat_warning!("Theme '{theme}' is deprecated, using 'ansi' instead.");
                     return self.get_theme("ansi");
                 }
                 if !theme.is_empty() {
-                    bat_warning!("Unknown theme '{}', using default.", theme)
+                    bat_warning!("Unknown theme '{theme}', using default.")
                 }
                 self.get_theme_set()
-                    .get(self.fallback_theme.unwrap_or_else(Self::default_theme))
+                    .get(
+                        self.fallback_theme
+                            .unwrap_or_else(|| default_theme(ColorScheme::Dark)),
+                    )
                     .expect("something is very wrong if the default theme is missing")
             }
         }
@@ -259,7 +212,7 @@ impl HighlightingAssets {
         language: Option<&str>,
         input: &mut OpenedInput,
         mapping: &SyntaxMapping,
-    ) -> Result<SyntaxReferenceInSet> {
+    ) -> Result<SyntaxReferenceInSet<'_>> {
         if let Some(language) = language {
             let syntax_set = self.get_syntax_set()?;
             return syntax_set
@@ -291,14 +244,17 @@ impl HighlightingAssets {
     pub(crate) fn find_syntax_by_name(
         &self,
         syntax_name: &str,
-    ) -> Result<Option<SyntaxReferenceInSet>> {
+    ) -> Result<Option<SyntaxReferenceInSet<'_>>> {
         let syntax_set = self.get_syntax_set()?;
         Ok(syntax_set
             .find_syntax_by_name(syntax_name)
             .map(|syntax| SyntaxReferenceInSet { syntax, syntax_set }))
     }
 
-    fn find_syntax_by_extension(&self, e: Option<&OsStr>) -> Result<Option<SyntaxReferenceInSet>> {
+    fn find_syntax_by_extension(
+        &self,
+        e: Option<&OsStr>,
+    ) -> Result<Option<SyntaxReferenceInSet<'_>>> {
         let syntax_set = self.get_syntax_set()?;
         let extension = e.and_then(|x| x.to_str()).unwrap_or_default();
         Ok(syntax_set
@@ -306,11 +262,18 @@ impl HighlightingAssets {
             .map(|syntax| SyntaxReferenceInSet { syntax, syntax_set }))
     }
 
+    fn find_syntax_by_token(&self, token: &str) -> Result<Option<SyntaxReferenceInSet<'_>>> {
+        let syntax_set = self.get_syntax_set()?;
+        Ok(syntax_set
+            .find_syntax_by_token(token)
+            .map(|syntax| SyntaxReferenceInSet { syntax, syntax_set }))
+    }
+
     fn get_syntax_for_file_name(
         &self,
         file_name: &OsStr,
         ignored_suffixes: &IgnoredSuffixes,
-    ) -> Result<Option<SyntaxReferenceInSet>> {
+    ) -> Result<Option<SyntaxReferenceInSet<'_>>> {
         let mut syntax = self.find_syntax_by_extension(Some(file_name))?;
         if syntax.is_none() {
             syntax =
@@ -326,7 +289,7 @@ impl HighlightingAssets {
         &self,
         file_name: &OsStr,
         ignored_suffixes: &IgnoredSuffixes,
-    ) -> Result<Option<SyntaxReferenceInSet>> {
+    ) -> Result<Option<SyntaxReferenceInSet<'_>>> {
         let mut syntax = self.find_syntax_by_extension(Path::new(file_name).extension())?;
         if syntax.is_none() {
             syntax =
@@ -341,11 +304,15 @@ impl HighlightingAssets {
     fn get_first_line_syntax(
         &self,
         reader: &mut InputReader,
-    ) -> Result<Option<SyntaxReferenceInSet>> {
+    ) -> Result<Option<SyntaxReferenceInSet<'_>>> {
         let syntax_set = self.get_syntax_set()?;
         Ok(String::from_utf8(reader.first_line.clone())
             .ok()
-            .and_then(|l| syntax_set.find_syntax_by_first_line(&l))
+            .and_then(|l| {
+                // Strip UTF-8 BOM if present
+                let line = l.strip_prefix('\u{feff}').unwrap_or(&l);
+                syntax_set.find_syntax_by_first_line(line)
+            })
             .map(|syntax| SyntaxReferenceInSet { syntax, syntax_set }))
     }
 }
@@ -390,33 +357,12 @@ fn asset_from_cache<T: serde::de::DeserializeOwned>(
 ) -> Result<T> {
     let contents = fs::read(path).map_err(|_| {
         format!(
-            "Could not load cached {} '{}'",
-            description,
+            "Could not load cached {description} '{}'",
             path.to_string_lossy()
         )
     })?;
     asset_from_contents(&contents[..], description, compressed)
         .map_err(|_| format!("Could not parse cached {description}").into())
-}
-
-#[cfg(target_os = "macos")]
-fn macos_dark_mode_active() -> bool {
-    const PREFERENCES_FILE: &str = "Library/Preferences/.GlobalPreferences.plist";
-    const STYLE_KEY: &str = "AppleInterfaceStyle";
-
-    let preferences_file = home::home_dir()
-        .map(|home| home.join(PREFERENCES_FILE))
-        .expect("Could not get home directory");
-
-    match plist::Value::from_file(preferences_file).map(|file| file.into_dictionary()) {
-        Ok(Some(preferences)) => match preferences.get(STYLE_KEY).and_then(|val| val.as_string()) {
-            Some(value) => value == "Dark",
-            // If the key does not exist, then light theme is currently in use.
-            None => false,
-        },
-        // Unreachable, in theory. All macOS users have a home directory and preferences file setup.
-        Ok(None) | Err(_) => true,
-    }
 }
 
 #[cfg(test)]
@@ -437,7 +383,7 @@ mod tests {
         pub temp_dir: TempDir,
     }
 
-    impl<'a> SyntaxDetectionTest<'a> {
+    impl SyntaxDetectionTest<'_> {
         fn new() -> Self {
             SyntaxDetectionTest {
                 assets: HighlightingAssets::from_binary(),
@@ -596,6 +542,41 @@ mod tests {
         );
         assert_eq!(
             test.syntax_for_file_with_content("my_script", "<?php"),
+            "PHP"
+        );
+    }
+
+    #[test]
+    fn syntax_detection_first_line_with_utf8_bom() {
+        let test = SyntaxDetectionTest::new();
+
+        // Test that XML files are detected correctly even with UTF-8 BOM
+        // The BOM should be stripped before first-line syntax detection
+        let xml_with_bom = "\u{feff}<?xml version=\"1.0\" encoding=\"utf-8\"?>";
+        assert_eq!(
+            test.syntax_for_file_with_content("unknown_file", xml_with_bom),
+            "XML"
+        );
+
+        // Test the specific .csproj case mentioned in the issue
+        // Even if .csproj has extension mapping, this tests first-line fallback
+        let csproj_content_with_bom = "\u{feff}<?xml version=\"1.0\" encoding=\"utf-8\"?>\n<Project ToolsVersion=\"15.0\" xmlns=\"http://schemas.microsoft.com/developer/msbuild/2003\">";
+        assert_eq!(
+            test.syntax_for_file_with_content("test.csproj", csproj_content_with_bom),
+            "XML"
+        );
+
+        // Test that shell scripts are detected correctly even with UTF-8 BOM
+        let script_with_bom = "\u{feff}#!/bin/bash";
+        assert_eq!(
+            test.syntax_for_file_with_content("unknown_script", script_with_bom),
+            "Bourne Again Shell (bash)"
+        );
+
+        // Test that PHP files are detected correctly even with UTF-8 BOM
+        let php_with_bom = "\u{feff}<?php";
+        assert_eq!(
+            test.syntax_for_file_with_content("unknown_php", php_with_bom),
             "PHP"
         );
     }
