@@ -3,6 +3,8 @@
 mod app;
 mod assets;
 mod clap_app;
+#[cfg(feature = "application")]
+mod completions;
 mod config;
 mod directories;
 mod input;
@@ -14,6 +16,8 @@ use std::io::{BufReader, Write};
 use std::path::Path;
 use std::process;
 
+use bat::output::{OutputHandle, OutputType};
+use bat::theme::DetectColorScheme;
 use nu_ansi_term::Color::Green;
 use nu_ansi_term::Style;
 
@@ -30,12 +34,12 @@ use directories::PROJECT_DIRS;
 use globset::GlobMatcher;
 
 use bat::{
-    assets::HighlightingAssets,
     config::Config,
     controller::Controller,
     error::*,
     input::Input,
     style::{StyleComponent, StyleComponents},
+    theme::{color_scheme, default_theme, ColorScheme},
     MappingTarget, PagingMode,
 };
 
@@ -158,7 +162,7 @@ pub fn get_languages(config: &Config, cache_dir: &Path) -> Result<String> {
         };
 
         for lang in languages {
-            write!(result, "{:width$}{}", lang.name, separator, width = longest).ok();
+            write!(result, "{:width$}{separator}", lang.name, width = longest).ok();
 
             // Number of characters on this line so far, wrap before `desired_width`
             let mut num_chars = 0;
@@ -169,7 +173,7 @@ pub fn get_languages(config: &Config, cache_dir: &Path) -> Result<String> {
                 let new_chars = word.len() + comma_separator.len();
                 if num_chars + new_chars >= desired_width {
                     num_chars = 0;
-                    write!(result, "\n{:width$}{}", "", separator, width = longest).ok();
+                    write!(result, "\n{:width$}{separator}", "", width = longest).ok();
                 }
 
                 num_chars += new_chars;
@@ -189,7 +193,12 @@ fn theme_preview_file<'a>() -> Input<'a> {
     Input::from_reader(Box::new(BufReader::new(THEME_PREVIEW_DATA)))
 }
 
-pub fn list_themes(cfg: &Config, config_dir: &Path, cache_dir: &Path) -> Result<()> {
+pub fn list_themes(
+    cfg: &Config,
+    config_dir: &Path,
+    cache_dir: &Path,
+    detect_color_scheme: DetectColorScheme,
+) -> Result<()> {
     let assets = assets_from_cache_or_binary(cfg.use_custom_assets, cache_dir)?;
     let mut config = cfg.clone();
     let mut style = HashSet::new();
@@ -197,36 +206,45 @@ pub fn list_themes(cfg: &Config, config_dir: &Path, cache_dir: &Path) -> Result<
     config.language = Some("Rust");
     config.style_components = StyleComponents(style);
 
-    let stdout = io::stdout();
-    let mut stdout = stdout.lock();
+    let mut output_type =
+        OutputType::from_mode(config.paging_mode, config.wrapping_mode, config.pager)?;
+    let mut writer = output_type.handle()?;
 
-    let default_theme = HighlightingAssets::default_theme();
+    let default_theme_name = default_theme(color_scheme(detect_color_scheme).unwrap_or_default());
     for theme in assets.themes() {
-        let default_theme_info = if default_theme == theme {
+        let default_theme_info = if default_theme_name == theme {
             " (default)"
+        } else if default_theme(ColorScheme::Dark) == theme {
+            " (default dark)"
+        } else if default_theme(ColorScheme::Light) == theme {
+            " (default light)"
         } else {
             ""
         };
         if config.colored_output {
             writeln!(
-                stdout,
-                "Theme: {}{}\n",
+                writer,
+                "Theme: {}{default_theme_info}\n",
                 Style::new().bold().paint(theme.to_string()),
-                default_theme_info
             )?;
             config.theme = theme.to_string();
             Controller::new(&config, &assets)
-                .run(vec![theme_preview_file()], None)
+                .run(
+                    vec![theme_preview_file()],
+                    Some(OutputHandle::IoWrite(&mut writer)),
+                )
                 .ok();
-            writeln!(stdout)?;
+            writeln!(writer)?;
+        } else if config.loop_through {
+            writeln!(writer, "{theme}")?;
         } else {
-            writeln!(stdout, "{theme}{default_theme_info}")?;
+            writeln!(writer, "{theme}{default_theme_info}")?;
         }
     }
 
     if config.colored_output {
         writeln!(
-            stdout,
+            writer,
             "Further themes can be installed to '{}', \
             and are added to the cache with `bat cache --build`. \
             For more information, see:\n\n  \
@@ -337,6 +355,18 @@ fn run() -> Result<bool> {
         return Ok(true);
     }
 
+    #[cfg(feature = "application")]
+    if let Some(shell) = app.matches.get_one::<String>("completion") {
+        match shell.as_str() {
+            "bash" => println!("{}", completions::BASH_COMPLETION),
+            "fish" => println!("{}", completions::FISH_COMPLETION),
+            "ps1" => println!("{}", completions::PS1_COMPLETION),
+            "zsh" => println!("{}", completions::ZSH_COMPLETION),
+            _ => unreachable!("No completion for shell '{shell}' available."),
+        }
+        return Ok(true);
+    }
+
     match app.matches.subcommand() {
         Some(("cache", cache_matches)) => {
             // If there is a file named 'cache' in the current working directory,
@@ -371,7 +401,7 @@ fn run() -> Result<bool> {
                 };
                 run_controller(inputs, &plain_config, cache_dir)
             } else if app.matches.get_flag("list-themes") {
-                list_themes(&config, config_dir, cache_dir)?;
+                list_themes(&config, config_dir, cache_dir, DetectColorScheme::default())?;
                 Ok(true)
             } else if app.matches.get_flag("config-file") {
                 println!("{}", config_file().to_string_lossy());
