@@ -10,6 +10,7 @@ use crate::{
     error::Result,
     input,
     line_range::{HighlightedLineRanges, LineRange, LineRanges},
+    output::OutputHandle,
     style::StyleComponent,
     StripAnsiMode, SyntaxMapping, WrappingMode,
 };
@@ -245,7 +246,9 @@ impl<'a> PrettyPrinter<'a> {
         self
     }
 
-    /// Specify the highlighting theme
+    /// Specify the highlighting theme.
+    /// You can use [`crate::theme::theme`] to pick a theme based on user preferences
+    /// and the terminal's background color.
     pub fn theme(&mut self, theme: impl AsRef<str>) -> &mut Self {
         self.config.theme = theme.as_ref().to_owned();
         self
@@ -279,6 +282,11 @@ impl<'a> PrettyPrinter<'a> {
     /// If you want to call 'print' multiple times, you have to call the appropriate
     /// input_* methods again.
     pub fn print(&mut self) -> Result<bool> {
+        self.print_with_writer(None::<&mut dyn std::fmt::Write>)
+    }
+
+    /// Pretty-print all specified inputs to a specified writer.
+    pub fn print_with_writer<W: std::fmt::Write>(&mut self, writer: Option<W>) -> Result<bool> {
         let highlight_lines = std::mem::take(&mut self.highlighted_lines);
         self.config.highlighted_lines = HighlightedLineRanges(LineRanges::from(highlight_lines));
         self.config.term_width = self
@@ -315,7 +323,16 @@ impl<'a> PrettyPrinter<'a> {
 
         // Run the controller
         let controller = Controller::new(&self.config, &self.assets);
-        controller.run(inputs.into_iter().map(|i| i.into()).collect(), None)
+
+        // If writer is provided, pass it to the controller, otherwise pass None
+        if let Some(mut w) = writer {
+            controller.run(
+                inputs.into_iter().map(|i| i.into()).collect(),
+                Some(OutputHandle::FmtWrite(&mut w)),
+            )
+        } else {
+            controller.run(inputs.into_iter().map(|i| i.into()).collect(), None)
+        }
     }
 }
 
