@@ -9,7 +9,6 @@ use tempfile::tempdir;
 mod unix {
     pub use std::fs::File;
     pub use std::io::{self, Write};
-    pub use std::os::unix::io::FromRawFd;
     pub use std::path::PathBuf;
     pub use std::process::Stdio;
     pub use std::thread;
@@ -164,7 +163,7 @@ fn line_numbers() {
         .arg("--decorations=always")
         .assert()
         .success()
-        .stdout("   1 line 1\n   2 line 2\n   3 line 3\n   4 line 4\n");
+        .stdout("   1 line 1\n   2 line 2\n   3 line 3\n   4 line 4\n   5 line 5\n   6 line 6\n   7 line 7\n   8 line 8\n   9 line 9\n  10 line 10\n");
 }
 
 #[test]
@@ -175,6 +174,46 @@ fn line_range_2_3() {
         .assert()
         .success()
         .stdout("line 2\nline 3\n");
+}
+
+#[test]
+fn line_range_up_to_2_from_back() {
+    bat()
+        .arg("multiline.txt")
+        .arg("--line-range=:-2")
+        .assert()
+        .success()
+        .stdout("line 1\nline 2\nline 3\nline 4\nline 5\nline 6\nline 7\nline 8\n");
+}
+
+#[test]
+fn line_range_up_to_2_from_back_single_line_is_empty() {
+    bat()
+        .arg("single-line.txt")
+        .arg("--line-range=:-2")
+        .assert()
+        .success()
+        .stdout("");
+}
+
+#[test]
+fn line_range_from_back_last_two() {
+    bat()
+        .arg("multiline.txt")
+        .arg("--line-range=-2:")
+        .assert()
+        .success()
+        .stdout("line 9\nline 10\n");
+}
+
+#[test]
+fn line_range_from_back_last_two_single_line() {
+    bat()
+        .arg("single-line.txt")
+        .arg("--line-range=-2:")
+        .assert()
+        .success()
+        .stdout("Single Line");
 }
 
 #[test]
@@ -191,10 +230,10 @@ fn line_range_first_two() {
 fn line_range_last_3() {
     bat()
         .arg("multiline.txt")
-        .arg("--line-range=2:")
+        .arg("--line-range=8:")
         .assert()
         .success()
-        .stdout("line 2\nline 3\nline 4\n");
+        .stdout("line 8\nline 9\nline 10\n");
 }
 
 #[test]
@@ -206,6 +245,137 @@ fn line_range_multiple() {
         .assert()
         .success()
         .stdout("line 1\nline 2\nline 4\n");
+}
+
+#[test]
+fn line_range_multiple_with_context() {
+    bat()
+        .arg("multiline.txt")
+        .arg("--line-range=2::1")
+        .arg("--line-range=8::1")
+        .assert()
+        .success()
+        .stdout("line 1\nline 2\nline 3\nline 7\nline 8\nline 9\n");
+}
+
+#[test]
+fn line_range_context_around_single_line() {
+    bat()
+        .arg("multiline.txt")
+        .arg("--line-range=5::2")
+        .assert()
+        .success()
+        .stdout("line 3\nline 4\nline 5\nline 6\nline 7\n");
+}
+
+#[test]
+fn line_range_context_around_single_line_minimal() {
+    bat()
+        .arg("multiline.txt")
+        .arg("--line-range=5::1")
+        .assert()
+        .success()
+        .stdout("line 4\nline 5\nline 6\n");
+}
+
+#[test]
+fn line_range_context_around_range() {
+    bat()
+        .arg("multiline.txt")
+        .arg("--line-range=4:6:2")
+        .assert()
+        .success()
+        .stdout("line 2\nline 3\nline 4\nline 5\nline 6\nline 7\nline 8\n");
+}
+
+#[test]
+fn line_range_context_at_file_boundaries() {
+    bat()
+        .arg("multiline.txt")
+        .arg("--line-range=1::2")
+        .assert()
+        .success()
+        .stdout("line 1\nline 2\nline 3\n");
+}
+
+#[test]
+fn line_range_context_at_end_of_file() {
+    bat()
+        .arg("multiline.txt")
+        .arg("--line-range=10::2")
+        .assert()
+        .success()
+        .stdout("line 8\nline 9\nline 10\n");
+}
+
+#[test]
+fn line_range_context_zero() {
+    bat()
+        .arg("multiline.txt")
+        .arg("--line-range=5::0")
+        .assert()
+        .success()
+        .stdout("line 5\n");
+}
+
+#[test]
+fn line_range_context_negative_single_line() {
+    bat()
+        .arg("multiline.txt")
+        .arg("--line-range=5::-1")
+        .assert()
+        .failure()
+        .stderr(predicate::str::contains(
+            "Invalid context number in N::C format",
+        ));
+}
+
+#[test]
+fn line_range_context_negative_range() {
+    bat()
+        .arg("multiline.txt")
+        .arg("--line-range=5:6:-1")
+        .assert()
+        .failure()
+        .stderr(predicate::str::contains(
+            "Invalid context number in N:M:C format",
+        ));
+}
+
+#[test]
+fn line_range_context_non_numeric_single_line() {
+    bat()
+        .arg("multiline.txt")
+        .arg("--line-range=10::abc")
+        .assert()
+        .failure()
+        .stderr(predicate::str::contains(
+            "Invalid context number in N::C format",
+        ));
+}
+
+#[test]
+fn line_range_context_non_numeric_range() {
+    bat()
+        .arg("multiline.txt")
+        .arg("--line-range=10:12:xyz")
+        .assert()
+        .failure()
+        .stderr(predicate::str::contains(
+            "Invalid context number in N:M:C format",
+        ));
+}
+
+#[test]
+fn line_range_context_very_large() {
+    bat()
+        .arg("multiline.txt")
+        .arg("--line-range=10::999999")
+        .assert()
+        .success()
+        .stdout(
+            "line 1\nline 2\nline 3\nline 4\nline 5\nline 6\nline 7\nline 8\nline 9\nline 10\n",
+        );
 }
 
 #[test]
@@ -274,11 +444,8 @@ fn squeeze_limit_line_numbers() {
 
 #[test]
 fn list_themes_with_colors() {
-    #[cfg(target_os = "macos")]
-    let default_theme_chunk = "Monokai Extended Light\x1B[0m (default)";
-
-    #[cfg(not(target_os = "macos"))]
     let default_theme_chunk = "Monokai Extended\x1B[0m (default)";
+    let default_light_theme_chunk = "Monokai Extended Light\x1B[0m (default light)";
 
     bat()
         .arg("--color=always")
@@ -287,34 +454,59 @@ fn list_themes_with_colors() {
         .success()
         .stdout(predicate::str::contains("DarkNeon").normalize())
         .stdout(predicate::str::contains(default_theme_chunk).normalize())
+        .stdout(predicate::str::contains(default_light_theme_chunk).normalize())
         .stdout(predicate::str::contains("Output the square of a number.").normalize());
 }
 
 #[test]
 fn list_themes_without_colors() {
-    #[cfg(target_os = "macos")]
-    let default_theme_chunk = "Monokai Extended Light (default)";
-
-    #[cfg(not(target_os = "macos"))]
     let default_theme_chunk = "Monokai Extended (default)";
+    let default_light_theme_chunk = "Monokai Extended Light (default light)";
 
     bat()
         .arg("--color=never")
+        .arg("--decorations=always") // trick bat into setting `Config::loop_through` to false
         .arg("--list-themes")
         .assert()
         .success()
         .stdout(predicate::str::contains("DarkNeon").normalize())
-        .stdout(predicate::str::contains(default_theme_chunk).normalize());
+        .stdout(predicate::str::contains(default_theme_chunk).normalize())
+        .stdout(predicate::str::contains(default_light_theme_chunk).normalize());
 }
 
 #[test]
-#[cfg_attr(any(not(feature = "git"), target_os = "windows"), ignore)]
+fn list_themes_to_piped_output() {
+    bat().arg("--list-themes").assert().success().stdout(
+        predicate::str::contains("(default)")
+            .not()
+            .and(predicate::str::contains("(default light)").not())
+            .and(predicate::str::contains("(default dark)").not()),
+    );
+}
+
+#[test]
+fn list_languages() {
+    bat()
+        .arg("--list-languages")
+        .assert()
+        .success()
+        .stdout(predicate::str::contains("Rust").normalize());
+}
+
+#[test]
+#[cfg_attr(
+    any(not(feature = "git"), feature = "lessopen", target_os = "windows"),
+    ignore
+)]
 fn short_help() {
     test_help("-h", "../doc/short-help.txt");
 }
 
 #[test]
-#[cfg_attr(any(not(feature = "git"), target_os = "windows"), ignore)]
+#[cfg_attr(
+    any(not(feature = "git"), feature = "lessopen", target_os = "windows"),
+    ignore
+)]
 fn long_help() {
     test_help("--help", "../doc/long-help.txt");
 }
@@ -400,14 +592,25 @@ fn stdin_to_stdout_cycle() -> io::Result<()> {
 
 #[cfg(unix)]
 #[test]
+fn bat_error_to_stderr() {
+    bat()
+        .arg("/tmp")
+        .assert()
+        .failure()
+        .stderr(predicate::str::contains("[bat error]"));
+}
+
+#[cfg(unix)]
+#[test]
 fn no_args_doesnt_break() {
     // To simulate bat getting started from the shell, a process is created with stdin and stdout
     // as the slave end of a pseudo terminal. Although both point to the same "file", bat should
     // not exit, because in this case it is safe to read and write to the same fd, which is why
     // this test exists.
+
     let OpenptyResult { master, slave } = openpty(None, None).expect("Couldn't open pty.");
-    let mut master = unsafe { File::from_raw_fd(master) };
-    let stdin_file = unsafe { File::from_raw_fd(slave) };
+    let mut master = File::from(master);
+    let stdin_file = File::from(slave);
     let stdout_file = stdin_file.try_clone().unwrap();
     let stdin = Stdio::from(stdin_file);
     let stdout = Stdio::from(stdout_file);
@@ -415,6 +618,7 @@ fn no_args_doesnt_break() {
     let mut child = bat_raw_command()
         .stdin(stdin)
         .stdout(stdout)
+        .env("TERM", "dumb") // Suppresses color detection
         .spawn()
         .expect("Failed to start.");
 
@@ -1011,6 +1215,31 @@ fn enable_pager_if_pp_flag_comes_before_paging() {
 }
 
 #[test]
+fn paging_does_not_override_simple_plain() {
+    bat()
+        .env("PAGER", "echo pager-output")
+        .arg("--decorations=always")
+        .arg("--plain")
+        .arg("--paging=never")
+        .arg("test.txt")
+        .assert()
+        .success()
+        .stdout(predicate::eq("hello world\n"));
+}
+
+#[test]
+fn simple_plain_does_not_override_paging() {
+    bat()
+        .env("PAGER", "echo pager-output")
+        .arg("--paging=always")
+        .arg("--plain")
+        .arg("test.txt")
+        .assert()
+        .success()
+        .stdout(predicate::eq("pager-output\n"));
+}
+
+#[test]
 fn pager_failed_to_parse() {
     bat()
         .env("BAT_PAGER", "mismatched-quotes 'a")
@@ -1223,6 +1452,38 @@ fn utf16() {
         .assert()
         .success()
         .stdout("hello world\n");
+
+    bat()
+        .arg("--plain")
+        .arg("--decorations=always")
+        .arg("test_UTF-16BE.txt")
+        .assert()
+        .success()
+        .stdout("hello world\nthis is a test\n");
+}
+
+#[test]
+fn utf16le() {
+    bat()
+        .arg("--decorations=always")
+        .arg("--style=numbers")
+        .arg("--color=never")
+        .arg("test_UTF-16LE-complicated.txt")
+        .assert()
+        .success()
+        .stdout("   1 上一伊刀\n   2 foo bar\n   3 hello world\n");
+}
+
+#[test]
+fn utf16be() {
+    bat()
+        .arg("--decorations=always")
+        .arg("--style=numbers")
+        .arg("--color=never")
+        .arg("test_UTF-16BE-complicated.txt")
+        .assert()
+        .success()
+        .stdout("   1 上一伊刀\n   2 foo bar\n   3 hello world\n");
 }
 
 // Regression test for https://github.com/sharkdp/bat/issues/1922
@@ -1393,6 +1654,12 @@ fn snip() {
    2 line 2
  ...─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ 8< ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─ ─
    4 line 4
+   5 line 5
+   6 line 6
+   7 line 7
+   8 line 8
+   9 line 9
+  10 line 10
 ",
         );
 }
@@ -1558,6 +1825,17 @@ oken
 ──────────
 ",
         )
+        .stderr("");
+}
+
+#[test]
+fn header_narrow_terminal_with_multibyte_chars() {
+    bat()
+        .arg("--terminal-width=30")
+        .arg("--decorations=always")
+        .arg("test.A—B가")
+        .assert()
+        .success()
         .stderr("");
 }
 
@@ -1793,7 +2071,7 @@ fn do_not_panic_regression_tests() {
     ] {
         bat()
             .arg("--color=always")
-            .arg(&format!("regression_tests/{filename}"))
+            .arg(format!("regression_tests/{filename}"))
             .assert()
             .success();
     }
@@ -1806,7 +2084,7 @@ fn do_not_detect_different_syntax_for_stdin_and_files() {
     let cmd_for_file = bat()
         .arg("--color=always")
         .arg("--map-syntax=*.js:Markdown")
-        .arg(&format!("--file-name={file}"))
+        .arg(format!("--file-name={file}"))
         .arg("--style=plain")
         .arg(file)
         .assert()
@@ -1816,7 +2094,7 @@ fn do_not_detect_different_syntax_for_stdin_and_files() {
         .arg("--color=always")
         .arg("--map-syntax=*.js:Markdown")
         .arg("--style=plain")
-        .arg(&format!("--file-name={file}"))
+        .arg(format!("--file-name={file}"))
         .pipe_stdin(Path::new(EXAMPLES_DIR).join(file))
         .unwrap()
         .assert()
@@ -1835,7 +2113,7 @@ fn no_first_line_fallback_when_mapping_to_invalid_syntax() {
     bat()
         .arg("--color=always")
         .arg("--map-syntax=*.invalid-syntax:InvalidSyntax")
-        .arg(&format!("--file-name={file}"))
+        .arg(format!("--file-name={file}"))
         .arg("--style=plain")
         .arg(file)
         .assert()
@@ -1926,6 +2204,16 @@ fn show_all_with_unicode() {
         .arg("control_characters.txt")
         .assert()
         .stdout("␀␁␂␃␄␅␆␇␈├─┤␊\n␋␌␍␎␏␐␑␒␓␔␕␖␗␘␙␚␛␜␝␞␟␡")
+        .stderr("");
+}
+
+#[test]
+fn binary_as_text() {
+    bat()
+        .arg("--binary=as-text")
+        .arg("control_characters.txt")
+        .assert()
+        .stdout("\x00\x01\x02\x03\x04\x05\x06\x07\x08\x09\x0A\x0B\x0C\x0D\x0E\x0F\x10\x11\x12\x13\x14\x15\x16\x17\x18\x19\x1A\x1B\x1C\x1D\x1E\x1F\x7F")
         .stderr("");
 }
 
@@ -2218,6 +2506,46 @@ fn theme_arg_overrides_env_withconfig() {
 }
 
 #[test]
+fn theme_light_env_var_is_respected() {
+    bat()
+        .env("BAT_THEME_LIGHT", "Coldark-Cold")
+        .env("COLORTERM", "truecolor")
+        .arg("--theme=light")
+        .arg("--paging=never")
+        .arg("--color=never")
+        .arg("--terminal-width=80")
+        .arg("--wrap=never")
+        .arg("--decorations=always")
+        .arg("--style=plain")
+        .arg("--highlight-line=1")
+        .write_stdin("Lorem Ipsum")
+        .assert()
+        .success()
+        .stdout("\x1B[48;2;208;218;231mLorem Ipsum\x1B[0m")
+        .stderr("");
+}
+
+#[test]
+fn theme_dark_env_var_is_respected() {
+    bat()
+        .env("BAT_THEME_DARK", "Coldark-Dark")
+        .env("COLORTERM", "truecolor")
+        .arg("--theme=dark")
+        .arg("--paging=never")
+        .arg("--color=never")
+        .arg("--terminal-width=80")
+        .arg("--wrap=never")
+        .arg("--decorations=always")
+        .arg("--style=plain")
+        .arg("--highlight-line=1")
+        .write_stdin("Lorem Ipsum")
+        .assert()
+        .success()
+        .stdout("\x1B[48;2;33;48;67mLorem Ipsum\x1B[0m")
+        .stderr("");
+}
+
+#[test]
 fn theme_env_overrides_config() {
     bat_with_config()
         .env("BAT_CONFIG_PATH", "bat-theme.conf")
@@ -2395,7 +2723,6 @@ fn lessopen_stdin_piped() {
 #[cfg(unix)] // Expected output assumed that tests are run on a Unix-like system
 #[cfg(feature = "lessopen")]
 #[test]
-#[serial] // Randomly fails otherwise
 fn lessopen_and_lessclose_file_temp() {
     // This is mainly to test that $LESSCLOSE gets passed the correct file paths
     // In this case, the original file and the temporary file returned by $LESSOPEN
@@ -2413,7 +2740,6 @@ fn lessopen_and_lessclose_file_temp() {
 #[cfg(unix)] // Expected output assumed that tests are run on a Unix-like system
 #[cfg(feature = "lessopen")]
 #[test]
-#[serial] // Randomly fails otherwise
 fn lessopen_and_lessclose_file_piped() {
     // This is mainly to test that $LESSCLOSE gets passed the correct file paths
     // In these cases, the original file and a dash
@@ -2440,8 +2766,6 @@ fn lessopen_and_lessclose_file_piped() {
 #[cfg(unix)] // Expected output assumed that tests are run on a Unix-like system
 #[cfg(feature = "lessopen")]
 #[test]
-#[serial] // Randomly fails otherwise
-#[ignore = "randomly failing on some systems"]
 fn lessopen_and_lessclose_stdin_temp() {
     // This is mainly to test that $LESSCLOSE gets passed the correct file paths
     // In this case, a dash and the temporary file returned by $LESSOPEN
@@ -2459,7 +2783,6 @@ fn lessopen_and_lessclose_stdin_temp() {
 #[cfg(unix)] // Expected output assumed that tests are run on a Unix-like system
 #[cfg(feature = "lessopen")]
 #[test]
-#[serial] // Randomly fails otherwise
 fn lessopen_and_lessclose_stdin_piped() {
     // This is mainly to test that $LESSCLOSE gets passed the correct file paths
     // In these cases, two dashes
@@ -2668,6 +2991,27 @@ fn highlighting_independant_from_map_syntax_case() {
 }
 
 #[test]
+fn map_syntax_target_syntax_case_insensitive() {
+    let expected = bat()
+        .arg("-f")
+        .arg("--map-syntax=*.config:json")
+        .arg("map-syntax_case.Config")
+        .assert()
+        .get_output()
+        .stdout
+        .clone();
+
+    bat()
+        .arg("-f")
+        .arg("--map-syntax=*.config:json")
+        .arg("map-syntax_case.Config")
+        .assert()
+        .success()
+        .stdout(expected)
+        .stderr("");
+}
+
+#[test]
 fn strip_ansi_always_strips_ansi() {
     bat()
         .arg("--style=plain")
@@ -2809,4 +3153,72 @@ fn strip_ansi_auto_does_not_strip_ansi_when_plain_text_by_option() {
     .expect("valid utf8");
 
     assert!(output.contains("\x1B[33mYellow"))
+}
+
+// Tests that style components can be removed with `-component`.
+#[test]
+fn style_components_can_be_removed() {
+    bat()
+        .arg({
+            #[cfg(not(feature = "git"))]
+            {
+                "--style=full,-grid"
+            }
+            #[cfg(feature = "git")]
+            {
+                "--style=full,-grid,-changes"
+            }
+        })
+        .arg("--decorations=always")
+        .arg("--color=never")
+        .write_stdin("test")
+        .assert()
+        .success()
+        .stdout("     STDIN\n     Size: -\n   1 test\n")
+        .stderr("");
+}
+
+// Tests that style components are chosen based on the rightmost `--style` argument.
+#[test]
+fn style_components_can_be_overidden() {
+    bat()
+        .arg("--style=full")
+        .arg("--style=header,numbers")
+        .arg("--decorations=always")
+        .arg("--color=never")
+        .write_stdin("test")
+        .assert()
+        .success()
+        .stdout("     STDIN\n   1 test\n")
+        .stderr("");
+}
+
+// Tests that style components can be merged across multiple `--style` arguments.
+#[test]
+fn style_components_will_merge() {
+    bat()
+        .arg("--style=header,grid")
+        .arg("--style=-grid,+numbers")
+        .arg("--decorations=always")
+        .arg("--color=never")
+        .write_stdin("test")
+        .assert()
+        .success()
+        .stdout("     STDIN\n   1 test\n")
+        .stderr("");
+}
+
+// Tests that style components can be merged with the `BAT_STYLE` environment variable.
+#[test]
+fn style_components_will_merge_with_env_var() {
+    bat()
+        .env("BAT_STYLE", "header,grid")
+        .arg("--style=-grid,+numbers")
+        .arg("--decorations=always")
+        .arg("--color=never")
+        .write_stdin("test")
+        .assert()
+        .success()
+        .stdout("     STDIN\n   1 test\n")
+        .stderr("");
 }
